exported the getAge function to a new util file to keep the code cleaner, matched the frontend age validation with the API specific date calc
